Use only one screener config file for v8 and v9
